requirements.txt
aiohttp==3.7.3
amqp==5.0.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
Django==3.1.3
django-celery-beat==2.1.0
django-classy-tags==2.0.0
django-cookie-law==2.0.4
django-cors-headers==3.7.0
django-environ==0.8.1
django-timezone-field==4.0
djangorestframework==3.12.2
idna==2.10
Jinja2==2.11.2
kombu==5.0.2
lxml==4.6.2
MarkupSafe==1.1.1
multidict==5.1.0
mysqlclient==2.0.1
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.2.2
yarl==1.6.3
python-dotenv